% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_relationships.R
\name{annotateIonTbl}
\alias{annotateIonTbl}
\title{Annotate Features Based on Ion Relationships}
\usage{
annotateIonTbl(tbl, pw)
}
\arguments{
\item{tbl}{Feature table (with mz, id, etc.)}

\item{pw}{Pairwise ion table from calculatePairWiseIons}
}
\value{
tbl with updated 'annotation' and 'direction'
}
\description{
Adds adduct, neutral loss, and ion-pair annotations to features.
}
