% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_relationships.R
\name{calculatePairWiseIons}
\alias{calculatePairWiseIons}
\title{Calculate Pairwise Ion Relationships}
\usage{
calculatePairWiseIons(mzs, rt, ids, adduct_data)
}
\arguments{
\item{mzs}{Numeric vector of m/z values}

\item{rt}{Retention time (shared across all mzs in this set)}

\item{ids}{Vector of feature IDs corresponding to mzs}

\item{adduct_data}{Output from prepareAdductData(), table of common positive 
adduct molecular weights and neutral losses for ms1 feature relationships}
}
\value{
Data.frame of inferred pairwise relationships
}
\description{
Examines m/z differences among features to infer potential charge states,
neutral losses, and adduct relationships.
}
