% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{graphMinalemine}
\alias{graphMinalemine}
\title{Run a test of neutral loss analysis using example MGF files}
\usage{
graphMinalemine()
}
\value{
No return value. Called for side effect of writing a GML file.
}
\description{
This function locates `.mgf` files in the `extdata` folder of the `mypackage`
package and runs them through `NeutralLossGraph::networkMGFs()` to generate a network.
}
\details{
This is a helper function intended for testing or demonstrating the
`networkMGFs` function in the `NeutralLossGraph` package. The generated GML
output is written to `package_test` in current working directory.
}
\examples{
graphMinalemine()
}
