% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{networkMGFs}
\alias{networkMGFs}
\title{Generate a Molecular Similarity Network from MGF Files}
\usage{
networkMGFs(
  files,
  cutoff = 0.5,
  outputGML = "msms_network",
  includeMS1 = FALSE,
  includeFragments = FALSE,
  rt_h = 30,
  mz_h = 0.005
)
}
\arguments{
\item{files}{A character vector of paths to MGF files.}

\item{cutoff}{Numeric threshold for pairwise similarity. Only spectra with a similarity
above this threshold are included in the network. Default is 0.5.}

\item{outputGML}{The file path prefix for saving the output GraphML file. Default is
`'msms_network'` in current working directory.}

\item{includeMS1}{Logical. If TRUE, directional MS1-based annotations are used to modify
edge weights in the network. Annotations will be added to network edges. Default is FALSE.}

\item{includeFragments}{Logical. If TRUE, pairwise calculations of two features will 
include fragment values as well as neutral losses. Default is FALSE.}

\item{rt_h}{Numeric. Clustering height for retention time grouping. Default = 30.}

\item{mz_h}{Numeric. Clustering height for m/z grouping. Default = 0.005.}
}
\value{
No return value. Writes a GraphML file to the specified path.
}
\description{
This function processes one or more MGF files and builds a molecular network based on
precursor and fragment mass spectra similarity. It filters, scores, annotates, and clusters
MS/MS spectra, then exports the result as a GraphML file for visualization.
}
\details{
This function computes pairwise similarity between MS/MS spectra based on fragment
mass-to-charge (m/z) values. It also incorporates MS1 annotation information if
specified, modifying edge weights for more meaningful clustering. The final result
is a similarity network stored in the GraphML format, which can be visualized using
tools such as Cytoscape.
}
\examples{
# Specify paths to MGF files and generate a similarity network
files <- list.files("data/mgf/", pattern = "\\\\.mgf$", full.names = TRUE)
networkMGFs(files, cutoff = 0.6, outputGML = "results/network", includeMS1 = TRUE)

}
