% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{outputGraphML}
\alias{outputGraphML}
\title{Export a Similarity Network to GraphML Format}
\usage{
outputGraphML(
  tbl,
  pmzs,
  source,
  ioncount,
  annotation,
  rt,
  contract,
  graph_name = "new_graph",
  config = "{}",
  labels
)
}
\arguments{
\item{tbl}{A numeric matrix representing pairwise similarity scores between spectra.
Similarity values should be between 0 (no similarity) and 1 (identical spectra).}

\item{pmzs}{A vector of precursor m/z values corresponding to the spectra.}

\item{source}{A vector containing the source file names for each spectrum.}

\item{ioncount}{A vector representing the number of ions (peaks) in each spectrum.}

\item{annotation}{A vector of annotation strings for each spectrum.}

\item{rt}{A vector of retention times corresponding to each spectrum.}

\item{contract}{A vector indicating which vertices should be merged based on
equivalence relations.}

\item{graph_name}{The desired name for the output GraphML file. Default is `'new_graph'`.}

\item{config}{A JSON formatted object containing the creation parameters for the output GraphML file. 
Default is `{}`.}

\item{labels}{A vector of labels to assign to the name property of the vertex for display in cytoscape.
If label value is NA, the mean mz value of the contracted vertex is assigned to the name property.}

\item{direction}{A vector describing the ion relationships for each spectrum.}
}
\value{
No return value. Writes the generated similarity network as a GraphML file to
  the specified path.
}
\description{
This function takes a similarity matrix, spectral annotations, and metadata to
generate a molecular similarity network as a GraphML file, suitable for visualization
in tools like Cytoscape. The network is based on pairwise fragment similarity between
MS/MS spectra.
}
\details{
This function processes a similarity matrix and corresponding spectral metadata to
create a graph object. It then exports the graph as a GraphML file, which is compatible
with popular network visualization tools like Cytoscape. The resulting network captures
the similarity between spectra based on fragment mass-to-charge ratios and MS1 annotations.
}
\examples{
# Example usage of outputGraphML function
outputGraphML(tbl = similarity_matrix, pmzs = precursor_mz, source = file_sources, 
              ioncount = ion_counts, annotation = annotations, 
              direction = directions, rt = retention_times, 
              contract = merge_indices, graph_name = "molecular_network")

}
